requirements.txt
flask
flask-cors
pandas
numpy
Pillow
scikit-learn
